pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "binaryvibes"
version = "0.1.0"
description = "AI-native framework for direct binary manipulation — bypassing source code to work at the machine level"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "capstone>=5.0",
    "keystone-engine>=0.9",
    "lief>=0.15",
    "unicorn>=2.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
]

[project.scripts]
bv = "binaryvibes.cli.main:cli"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
